[js_fuzzer] Add allocation-timeout mutator
This adds a mutator that sprays %SetAllocationTimeout calls to
generate richer GC coverage.

The mutator is added in the end of the mutation chain, even after
try-catch mutations. It's not needed to run any other mutator after
this as it might only lead to poor mutations (e.g. the arguments
to SetAllocationTimeout should not get re-fuzzed).

The mutator is enabled in 50% of the cases and mutates in 2% of the
call sites. We boost the probability a bit before function calls
and right after %Optimize{Function|Maglev}OnNextCall calls.

Similar to GC, we count the calls in a library function and restrict
them to an upper limit to avoid poor performance if the function
gets called in loops.

The added tests run only the extreme paths, i.e. mutation
fully chosen or not chosen at all.

As a fly-by this adds a library method for zero settings which was
code duplicated among several tests already.

Bug: 417413668
Change-Id: I69cfabd19bc3cce487b0b50d11a1c5d052c74128
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6553949
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100374}