[ia32][arm] Decode param count correctly in kArchPrepareCallCFunction
The parameter count got encoded as a single number in the MiscField
BitField encoding, but it got decoded as two numbers, one in the
ParamField and one in the FPParamField encoding. With this CL the
parameter count gets decoded as a single number in the MiscField.

R=thibaudm@chromium.org

Change-Id: I0f0754a266a1ec360b37e57a2916875bb6b55473
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6563865
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100341}