[sandbox] Harden TailCalls after code update
Instead of letting CompileLazy and friends return the new code object
always reload it from the dispatch handle (or closure).

This ensures that whatever confusion can be caused to the runtime
function we always call a code object with a parameter count that
matches the initially checked one.

Bug: 40931165
Fixed: 417636716

Change-Id: I7c70e178f6631a587772c55a90c511ded89fb466
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6554252
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100345}