[wasm][fuzzer] Fix near-heap-limit callback
We are not allowed to directly throw an exception there. Trying to
support this during heap allocation (which triggers GC, which runs the
callback) seems very involved, so instead we fix this by just setting a
termination request which will be served at the next stack check.

This is in line with the existing comment two lines below the change.

R=jkummerow@chromium.org

Fixed: 406925416
Change-Id: Ib623f7c9947c3ec1f7bb10f01e6bebda0eaf0e0e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6563344
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100378}