[wasm] Wasm code coverage: Add BasicBlockCalculator
Spec:
https://docs.google.com/document/d/1NaD50lRija9DxHfb8X_rRlJ8EriqG_R20GM9E8Xf9ig/edit?pli=1&tab=t.0

This is the first in a series of CLs to add support for WebAssembly code
coverage in V8.
This initial CL adds code for a `BasicBlockCalculator`, which divides
the code of a Wasm function in a sequence of basic blocks. This will be
later used to add coverage instrumentation in Liftoff and
Turbofan/Turboshaft.

Bug: 354020953
Change-Id: Ieccec6ccb612a25b5007f284031b995e6044486c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6518896
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100384}