DescriptorArray: Never override slow iterable state
When overriding a key, we would previously set the fast iterable state
to "Unknown". However this is not correct, as during JSON.parse we might
generalize the descriptor during recursive parsing of the value, which
then makes us believe that all keys are fast iterable when we are done
with the original object.
To prevent this issue, we never override the fast iterable state if it
was known to be slow and only reset it when replacing a key if it was
previously fast.

Fixed: 416913840
Change-Id: I6906f335cfd8bea8c261ee42a23eb54e70bc9c42
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6565350
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100389}