Add two API calls for sandbox users.
SandboxContains is useful for asserts for embedders
who want to be sure their ArrayBuffer backings are
allocated in the correct sandbox (for those using
multiple IsolateGroups and therefore multiple sandboxes).

For embedders providing a custom ArrayBufferAllocator they
need access to the virtual address space so they can get
in-sandbox pages for their allocator.  There was already
an API method GetSandboxAddressSpace(), but this uses the
current group/sandbox, which is inconvenient when setting
up multiple isolate groups/sandboxes.

Change-Id: I94f1fcb80b5ca76df038928decc06887993682a9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6575561
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100420}