Remove last uses of GetIsolate and GetHeap
This is the manual follow-up to the mostly automatic
https://crrev.com/c/6495413.

Those methods will be removed in a follow-up
(https://crrev.com/c/6582963).

Depending on what makes most sense, we either use `Isolate::Current()`
directly, or store it in a variable that is reused multiple times, or we
pass the isolate as a parameter.

R=leszeks@chromium.org

Bug: 396607238
Change-Id: I9366410849cd4fa39fb792db9bb2ad35dba86ffe
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6582961
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100468}