[turbo*, maglev] Make Runtime_TransitionElementsKind update AllocationSites
Maglev & Turbofan generate code which transitions arrays during
element loads, if the feedback for the element load is polymorphic.

This CL records that transition in the AllocationSite, so that
future arrays created with the same AllocationSite have the correct
ElementsKind from the start.

Bug: 412888237
Change-Id: Ifc700397386c3a04599ffc4b9aa5fca402ec8d6f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6519400
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100507}