[regexp] Extend emoji regexp optimization to v mode
The optimization targets big alternations like
/cat|dog|car|chat|mouse/ in a few steps.  First they are
sorted, so you get something like /car|cat|chat|dog|mouse/,
and then prefixes are rationalized so you get /c(?:ar|at|hat)|dog|mouse/.  The optimization was broken
in v mode because of the way RegExpText is used to wrap
RegExpAtom, which meant the pattern was not recognized
and the rationalization needed to be rewritten.

Be sure to benchmark with --no-regexp-results-cache.

A factor of 10 is achieved on the ReplaceAllEmoji_gv benchmark.

Bug: 40216006
Change-Id: I7a157c1b4004963ae7f55652c189aae13462ff11
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6586794
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100524}