[foozzie] Implement remaining parts for independent deep printing
This adds a stand-alone implementation for the remaining bits of
deep printing that previously fell back to the mjsunit version.

When running differential fuzzing with Fuzzilli samples, we don't
include mjsunit.js and can't use the fallback.

The new parts are similar to what we have in mjsunit.js, but in
some parts a bit simpler.

Bug: 419144823
Change-Id: I11341e451e02037d109a095417b14f69798c551d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6592391
Reviewed-by: Marja Hölttä <marja@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100553}