More fine grained speculation control in charAt builtins
Use a second bit for the SpeculationMode in call feedback. This allows
us to do more fine grained speculations. In this CL we use this to
both speculate on a charAt being called on a string and the char not
being out of bounds.

Bug: 420433007

Change-Id: I32e63a62891a63deaed139d0bde17c540a58c4d1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6582363
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100547}