[maglev] Escape analysis fixe for drop_use
If we remove a use we can't know if it was an escaping one or not. To
be safe we need to assume it was. Otherwise it could happen that we
decrement the use count to a non-escaping use count even though there
is still an escaping use left.

Change-Id: Ifab2b559247189aad5574b49b0b038b4796f4607
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6596535
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100560}