[ic] Keyed load/store: support "1 map, many string keys" mode
and use it for Wasm objects with prototype chains. The key goal is
to collect type feedback for this situation, so that the optimizing
compiler can inline accesses (in an upcoming patch).
This mode is implemented as a new variant of monomorphic state, the
handler for it is the existing generic stub.
JS objects don't use the new IC state, because going straight to the
generic stub (as we used to) is faster for them.

Bug: 403372470
Change-Id: I75997ffa8fefc94d19313c040309c4b0b74443d7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6594352
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100558}