[proxy] Distinguish revocable Proxies
Proxies created with the default "new Proxy" constructor are not
revocable, and that property is useful for building fast paths
for them. So this CL adds a field JSProxy::flags, which is used
to annotate Proxies as revocable when applicable. (No actual fast
paths are added in this patch.)

Bug: 403372470
Change-Id: Ia240eaeadb80369b393a555d9de727ffd727f874
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6593533
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100561}