[test] Run basic V8 tests with all arm64 cpu features on simulators
This runs the V8 tests in the default shipping variant with a flag
that simulates all cpu features. Shards are scaled accordingly.

We keep the release and debug bots in sync, since the release bot
in CQ runs with dchecks on.

Bug: 402796656, 420743936
Change-Id: Idff2cac099e2a8f31f4feefd550982f51f3b44d9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6597276
Auto-Submit: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100614}