[roots] Reorder roots according to trace-serializer
The first 32 root indices are intended to be the most common roots in
the startup snapshot, so that they can be encoded more efficiently.
Running snapshot generation with --trace-serializer discovered that this
wasn't the case in practice, likely due to changes in the read-only
space serialization.

This reorders the roots so that the top 32 serialized roots are either
in the first 32 root indices, or they are awkward to change (because
they're torque generated maps or well-known symbols in their own macro
lists). The remaining roots are alpha-sorted for convenience.

This reduces the size of the snapshot (both V8's and blink's) by around
1%.

Change-Id: Iefc676286ba409583ac380c97b4c2b4d4b6e0ff3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6607222
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100607}