[heap] Allow to disable heap limit check in CollectGarbage().
This CL adds an additional argument to CollectGarbage() which allows
to disable the heap limit check. The heap limit check in
CollectGarbage() is the location where most OOMs are reported. As a
first user CollectAllAvailableGarbage() makes use of this argument and performs the heap limit check once after the subsequent GCs are done.

Bug: 40270831
Change-Id: I8b66b1c6b4a0543a971f658d4e17be35c4870c9a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6596556
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100618}