[turbofan] Make InferMaps explicit about when writes are safe
InferMapsUnsafe treats opcodes that can do a write as potentially
invalidating maps, but only in its generic (default swtich) path; the
specific opcode handling has to remember to consider writes as
potentially map-writing if needed. This can lead to issues when adding
an opcode to the switch and forgetting that this opcode potentiall
writes.

Now we make ignoring writes an explicit decision, which is checked in
debug; an ignored writing opcode must be annotated as ignoring the
write, and otherwise it must have marked maps as potentially
invalidated.

Fixed: 408160281
Change-Id: Ia81838e42507b35747adbcafad69c0bd05f58289
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6616342
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100664}