[turbofan] Add CheckedInt64ToAdditiveSafeInteger
The representation changer was bailing out if the use info rep
was equal to output rep. But this is not correct if the use info
type check is AdditiveSafeInt (similat to BigInts).

This CL adds a check to int64 inputs to an additive safe int.

Bug: 384959125
Change-Id: I57ade6fa0899dcbc225d52e042c02e9cceea5baf
Fixed: 421801277
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6616882
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100668}