[heap] Retry GCs on heap limit reached in MinorGC tasks
There is a non-negligible number of "reached heap limit" OOM errors
in MinorGC tasks. This CL tries to fix this in a similar manner to how
we addressed the OOMs in the allocation failure code path in
https://crrev.com/c/6611725. Instead of crashing immediately when
the heap limit is reached, we perform more and stronger GCs until
either enough garbage was collected or CollectAllAvailableGarbage()
is reached. This way we should always reach the last resort GC before
crashing for this kind of OOMs.

Bug: 40270831
Change-Id: I199cba853219741b103bcc4dd8a5a0fdbab50fd0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6618873
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100678}