[temporal] Fill in some of PlainTime, fix up Instant and Duration
Fills in:

 - The rest of Instant and Duration except for the parts needing
   timezone stuff (needs an upstream release for ZonedDateTime bindings)
 - Some of PlainTime: construction, arithmetic, toString.

Also removes getISOFields(), these APIs are no longer in the spec.

Tries to better organize the code.

(sorry this got big, I was hoping for it to be a clean addition of APIs
but a lot of these changes triggered refactors in the existing code. As
we add more of these that should reduce.)

Change-Id: I180c297d41377966ec4dc29a27a6fabb216276f8
Bug: 401065166
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6619917
Auto-Submit: Manish Goregaokar <manishearth@google.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Manish Goregaokar <manishearth@google.com>
Cr-Commit-Position: refs/heads/main@{#100677}