[wasm][jspi] Reintroduce an SBXCHECK
Reintroduce an SBXCHECK to ensure that we always suspend/return to the
immediate parent stack. Returning should already be safe because the
StackMemory chain lives outside of the sandbox, but suspending is still
unsafe because it uses the in-sandbox WasmSuspenderObject chain which
can be corrupted.

This is a short-term fix. A more robust solution will be to move the
WasmSuspenderObjects outside of the sandbox, like we did with the
WasmContinuationObjects (https://crrev.com/c/6426203).

R=jkummerow@chromium.org

Fixed: chromium:422645418
Change-Id: I08a60303ff65a07ae0ae4bd2f1e8758b174fba1a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6623289
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100689}