[heap] Add large page pool
This CL extends the page pool to also cache large pages. Dead new
large objects are added to a page pool up to a default limit of 32M.
Subsequent large object allocations can make use of the pool. Such
allocations reuse the smallest large page sufficient in size for
the large object allocation. When this page is too large, the
following full GC will trim the size of the page to the object size.

Large pages remain in the pool by default for about 3 seconds. If
the page is unused until then, it is freed on a background thread.
When enabled, this improves Jetstream 2 on M1 by about 0.3% and
0.2% on Linux.

Bug: 416640737
Change-Id: I0b307f50286ff03c0bc3e4a51a1d7d21dbfb275b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6555762
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100708}