[wasm][fuzzer] Disable Liftoff code flushing
... during the reference execution. The reference execution uses special
Liftoff code which has instrumentation to track for the number of
instructions executed and check for nondeterminism. Flushing Liftoff
code will recompile that code and replace it by regular Liftoff code
without those special checks.

In order to have a nice place to put the flag scope (and avoid affecting
the non-reference execution) this CL refactors the code to have the
reference run in a separate function.
It also adds an early exit in case no main function is exported; this is
now checked on the decoded module instead of instantiating it first and
then looking up the export.

R=jkummerow@chromium.org

Bug: 374820219
Change-Id: Ief3b576cb1e0ad52b4fb8c6042f9ff72406b4cae
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6618187
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100704}