[fastcall] Return value after deopt after fast call is undefined
Even if the return type of an API function is `void`, there is still a
return value in JavaScript, undefined.

This CL hard-codes the return value in the assembly code of the
`DeoptimizationEntry_LazyAfterFastCall` builtin. Ideally the return
value is handled in a continuation builtin, as suggested in
crbug.com/418936518, but that turned out to be more difficult than
expected.

R=leszeks@chromium.org

Bug: 422099361
Change-Id: I2d98deeff6a27b51cff9b312246816982ddd7dd1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6623291
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100706}