[wasm][growable-stacks] Prevent stack growth on central stack
The "grow_stack" external C call returns 0 to indicate to the caller
that we cannot grow the stack if we are currently on the central stack.
However this only handled the case where we are on the initial central
stack segment, not when we have switched back to the central stack from
a secondary stack.

R=clemensb@chromium.org

Fixed: 421055121
Change-Id: I818a655e1a82225bfd5072152b79bfc5ce6dbaf0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6619502
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100738}