[json] Parse: Reload number of descriptors in each iteration
When parsing an object with fast/unknown iterable feedback, we need to
reload the number of descriptors from the DescriptorArray instead of
keeping it in a register. The reason is that ParseJsonPropertyValue()
can trigger a GC, which might trim the DescriptorArray.

Fixed: 423459708
Change-Id: I0204c9743e97850d85539a31b2d3512e6b3243b7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6632608
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100742}