[turbofan] Ensure we emit a deopting node if operation is unreachable
The additive operation in this test has valid inputs, but overflows.

Since the inputs are known in compile-time, we emit an unreachable
after this node.

If we have int32 truncation, we will emit a pure int32 operator
that does not check overflows. This CL forces to emit an
operation with overflow check if the type of the operation
is invalid.

Note that, normally, it is safe to not check the overflow on
truncation, since the inputs are additive safe ints, the output
must be a safe integer. Therefore it truncates safely to int32.

Fixed: 423674623
Change-Id: I3737ea575b918b35e6697cde0e014d68b1297c3a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6632613
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100746}