[wasm][jspi] Move suspender object to trusted space
WasmSuspenderObject contains sensitive fields, in particular a
reference to the parent Suspender object and an external pointer
to the StackMemory. On top of the SBXCHECK reintroduced in [1],
protect the suspender object itself by moving it to trusted space.

- WasmSuspenderObject#parent becomes a protected pointer
- WasmResumeData#suspender becomes a trusted pointer

This makes the suspender non comparable with undefined, so use Smi
0 instead to represent the end of the list of nested suspenders.

However using a Smi for the ActiveSuspender root causes issues with
the heap snapshot generator, which expects a heap object there. Remove
the root and use an IsolateData field instead.

[1] https://chromium-review.googlesource.com/c/v8/v8/+/6426203

R=jkummerow@chromium.org

Bug: 422645418
Change-Id: I90642625eee3d4a42d6e11ab0bf9489b3049d07d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6633356
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100804}