[sandbox] Add foundation for hardware sandbox support, part 1
This CL builds the foundation for (experimental) hardware sandboxing
support based on PKEYs:
* It refactors the existing SandboxHardwareSupport class to use three
  PKEYs: the sandbox_pkey for in-sandbox memory, the out_of_sandbox_pkey
  for memory outside the sandbox, and the extension_pkey for additional
  out-of-sandbox memory that must also be accessible to sandboxed code.
  The latter should be considered an exception and will ideally be
  removed entirely in the future.
* It implements EnterSandbox and ExitSandbox routines for C++ and in the
  MacroAssembler. While in sandboxed execution mode, write access to the
  out_of_sandbox_pkey is removed.

Ideally, the out_of_sandbox_pkey would simply be the default pkey (key
zero), in which case code executing in sandboxed mode would not be able
to write to any out-of-sandbox memory (except "extension" memory).
However, that needs further refactorings to make it work and it also
requires some workarounds on Linux. As such, for now out-of-sandbox
memory must manually be registered via RegisterOutOfSandboxMemory, which
will set the out_of_sandbox_pkey for that memory. This then allows for a
gradual development of the initial prototype as out-of-sandbox memory
can be marked as such incrementally. With this CL, no memory is marked
in this way. This will happen in follow-up CLs.

Bug: 350324877
Change-Id: I8b5b370fe34d8835b4dc2b1e0c28f3b0be4aa4a4
Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6520459
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100786}