[turboshaft] Add verification for LateLoadElimination
This CL introduces the --turboshaft_verify_load_elimination flag. When
true, it causes LateLoadElimination to insert verification code before
each elimination, which will lead to crashes when LateLoadElimination
fails.

This will make it easier to fuzz LateLoadElimination, since we've seen
that wrong elimination have a tendency to not cause crashes, which
leads to Fuzzilli not detecting them.

I've confirmed that this verification leads to crashes on the repros
in https://crrev.com/c/6564404 (well, when undoing the fix).

Bug: 404257893
Change-Id: Ie178c7bdeb57a46e06bf9835744daf8d21a03910
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6632920
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100858}