[sandbox] Move WasmImportData.wrapper_budget into the sandbox
We do not want to write to trusted memory from sandboxed code.
Previously, the generic WasmToJS wrapper was writing to the
WasmImportData.wrapper_budget field in trusted space. To avoid that,
this CL now makes that field a Cell, and thereby moves it into the
sandbox. This is consistent with the wrapper_budget field of
WasmExportedFunctionData, which is used for JSToWasm wrappers.

Bug: 350324877
Change-Id: I28b6e850d48c5430d96ec34318d6e150a7a37f8a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6641578
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100808}