[temporal] Add native slots for ZonedDateTime
This is like https://chromium-review.googlesource.com/c/v8/v8/+/6604681,
except for ZonedDateTime (we didn't have the temporal_capi bindings
for ZDT when that CL was made).

This lets us remove a lot of the legacy torque-based code.

I've tried to add stubs for some of the newer ZDT APIs
but there's still more left. I'll do that in a separate CL.

Bug: 401065166
Change-Id: I6a6a6964f369902008dd3a924dc688042cdc6444
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6634233
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Manish Goregaokar <manishearth@google.com>
Cr-Commit-Position: refs/heads/main@{#100852}