[runtime] Allow Cell to contain weak values
The Cell has got two groups for accessors - maybe_value and value.
The former is for accessing MaybeObject value and the latter is for
use cases when it's known that the value is not weak.

This CL makes the cleared weak reference a value indicating a valid
prototype validity cell value.

This is a preparation for storing weak references in valid prototype
validity cells.

Bug: 424154602
Change-Id: I6acacc95769a922b399dd7dafd4e4d13e19f9e00
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6641901
Reviewed-by: Patrick Thier <pthier@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100828}