[d8][test] Fix pushing of android resources
The script that pushes test resources to android devices is not happy
when a resource turns out to be a directory. This may happen if, e.g.,
a JavaScript file contains `import('.')`; this is probably meaningless
in JavaScript but should still not fail with an exception before the
test even starts.

This CL fixes the issue by ignoring directory resources. It also
re-enables a test that had been skipped on android, for this reason.

Bug: 424836763
Bug: 422706696
Change-Id: I419513b42fc874c95b202812a0cc8d22727ba3df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6650906
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100883}