[objects, heap] Introduce SafeHeapObjectSize as uint32_t alias
A lot of code in the GC uses unsigned sizes (e.g. size_t) to copy
around bytes or iterate objects. The size of on-heap objects is stored
as int.

Attackers can craft negative sizes that can lead to very large offsets
via sign extension and e.g. conversion to size_t.

We can mitigate this problem by relying on SafeHeapObjectSize which is
an alias to uint32_t. The alias allows for highlighting potentially
violations and incrementally changing code. uint32_t is safe as it can
always be used as byte size directly or cast to size_t without
exceeding the sandbox.

Bug: 425896305
Change-Id: I30abb716f7fa588f60991886d956a6024b170e2e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6651890
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100896}