[Wasm interpreter] Updating reference offset position on copy
There is a problem in the GenericWasmToJSInterpreterWrapper. When
calling a JS function that returns multiple results, and one of those
results is of Ref type, that result was copied into an incorrectly
location into the packed_array used to pass arguments and return values
for the Wasm->JS call. This can cause an OOB write.

This also fixes another potential issue in the same code, for F32
results, that are written as doubles, not floats, into the packed_array.
This can cause an OOB write if an F32 is the last result type.

Change-Id: I26dbc9e504434a1387cc42883ce714e897717cb7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6653066
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Emmanuel Romero Ruiz <emromero@microsoft.com>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100958}