[d8] Fix basic support for module import, part 2
Currently, d8 provides some basic support for importing modules, used
for testing purposes. This CL fixes a bug in its implementation, which
is triggered if module A is imported twice, and A imports B which does
not exist or is erroneous. Module namespace object caching should only
apply to modules that are loaded and linked successfully, therefore, in
the previous scenario, module A should not be cached.

Bug: 424617296
Change-Id: I3d89f984546843e413ab82d93c4483ddad49023c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6641214
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100890}