Implement JSIterator exhaustion for JSTypedArray
The spec indicates that once an array iterator has reached an exhausted
state, it cannot back to a non-exhausted state. [1]

However, the current V8 implementation does not check the exhaustion for
TypedArray. This leads to the iterator being able to become
non-exhausted state if the length size changes.

Therefore, add an execution path to set the value check next_index to
max value for TypedArray like other receivers do. And also implement in
js-call-reducer for ReduceArrayIteratorPrototypeNext.

[1] https://tc39.es/ecma262/#sec-%arrayiteratorprototype%.next

Fixed: 415354777
Change-Id: I8cdfc2e020c209de7d55d45cf549c22bb419ed86
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6471936
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Seokho Song <seokho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100997}