[riscv] Use byte-sized margin for trampoline pool checks
We already use a byte-sized margin for BlockPoolsScope and there is at
least one caller of CheckTrampolinePoolQuick that already passes the
size in bytes (see CodeGenerator::PrepareForDeoptimizationExits).

Under the assumption that we're working towards getting rid of
BlockTrampolinePoolFor and replace all uses of it with scopes, I've kept
the margin there as the number of instructions. If we are to keep it, I
suggest we transition to using byte-sized margins everywhere.

Change-Id: I537458e4a1e3668f9d3d1c49cfa93f6811caa9a7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6650984
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#100939}