[fastcall][deopt] Skip frame of fast caller when searching for handler
When an exception is thrown, `UnwindAndFindHandler` iterates over the
stack and tries to find an exception handler for the thrown exception.
This is done by loading the return addresses from the stack and
comparing them with the entries in the handler tables of the code
objects that correspond to the stack frames.

When lazy deopt is triggered, the return address on the stack gets
overwritten with the address of a deopt exit, so that after the function
return of the callee the execution does not continue after the call
instruction but at the deopt exit. To make exception handling work with
the modified return address on the stack, `UnwindAndFindHandler` calls a
translation function which checks if the return address is a deopt exit,
and if so, translates the return address back to the original value.

When a function does a fast API call, and in this call an exception is
thrown and the function gets marked for deoptimization, then
`UnwindAndFindHandler` is not called within the fast API call, but
within the builtin that gets called by the deopt exit. This means the
return address on the stack for the function neither points after the
fast API call nor at the deopt exit of the fast API call, but at the
instruction after the call in the deopt exit, which is the deopt exit
for the next deopt point.

Therefore, as the return address on the stack is neither the instruction
after the fast API call nor its deopt exit, `UnwindAndFindHandler` does
not look up the correct entry in the handler table and therefore may
behave incorrectly.

This CL fixes the issue by adjusting the stack iteration in
`UnwindAndFindHandler`. The frame of the function that did the fast API
call gets handled already anyways by the frame of the deoptimization
builtin. Therefore, when the frame of the builtin gets processed, it
sets a flag that the next frame, i.e. the frame of the caller function,
should be ignore in the stack iteration.

Bug: 424627229
Change-Id: I5997f4b85d82f41bbba0a750f6514a4b7d1db50c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6652839
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100942}